Extract AssistBridge collaborator
Moves the LLM-assist surface into a new categorization/assist.py:
RedactedTransaction (the frozen privacy-contract dataclass), the
_amount_sign_label helper, and the AssistBridge class with the
categorize_assist read. The bridge is fully self-contained — it does not
touch the matcher or applier — so the facade exposes categorize_assist as
a thin pass-through to self._assist.

test_categorize_assist_clamps_to_max_batch_size now patches
get_settings on moneybin.services.categorization.assist (where the call
actually happens after the move), not on the package surface, and the
test_amount_sign_label tests import the private helper from the assist
submodule directly rather than via the package re-export.

Completes the four-way split (facade + matcher + applier + assist) along
the seams documented in docs/specs/categorization-matching-mechanics.md
and docs/specs/categorization-cold-start.md. No behavior change; full
suite passes.
