+++ src/moneybin/services/categorization/__init__.py
@@ -18,7 +18,7 @@ from typing import Any, Literal
 
 import duckdb
 
-from moneybin.config import get_settings
+from moneybin.config import get_settings as get_settings
 from moneybin.database import Database
 from moneybin.errors import UserError as UserError
 from moneybin.metrics.registry import (
@@ -40,7 +40,9 @@ from moneybin.metrics.registry import (
 from moneybin.protocol.envelope import ResponseEnvelope, build_envelope
 from moneybin.services._text import (
     build_match_inputs,
-    redact_for_llm,
+)
+from moneybin.services._text import (
+    redact_for_llm as redact_for_llm,
 )
 from moneybin.services.audit_service import AuditService
 from moneybin.services.categorization._shared import (
@@ -82,6 +84,10 @@ from moneybin.services.categorization.applier import (
     MatchApplier,
     RuleCreationResult,
     WriteOutcome,
+)
+from moneybin.services.categorization.assist import (
+    AssistBridge,
+    RedactedTransaction,
 )
 from moneybin.services.categorization.matcher import (
     CategorizationMatcher,
@@ -133,41 +139,6 @@ class CategorizationStats:
                 "Use transactions_categorize_pending_list to see uncategorized transactions"
             ],
         )
-
-
-@dataclass(frozen=True)
-class RedactedTransaction:
-    """LLM-safe view of an uncategorized transaction.
-
-    Type-enforces the redaction contract: no full amount, no date, no account ID.
-    The v2 contract (per categorization-matching-mechanics.md §Match input) adds
-    memo and structural-field signals. Adding any new field requires conscious
-    code review — accidental PII leakage is a compile-time impossibility enforced
-    by the frozen dataclass shape.
-    """
-
-    transaction_id: str
-    description_redacted: str
-    memo_redacted: str
-    source_type: str
-    transaction_type: str | None
-    check_number: str | None
-    is_transfer: bool
-    transfer_pair_id: str | None
-    payment_channel: str | None
-    amount_sign: Literal["+", "-", "0"]
-
-
-def _amount_sign_label(amount: float | None) -> Literal["+", "-", "0"]:
-    """Map a raw amount to the LLM-facing sign signal.
-
-    ``"0"`` covers both ``NULL`` (defective import) and zero amount (balance
-    adjustments, voided rows). Mapping both to ``"+"`` biases the LLM toward
-    income-side categories on rows that are neither income nor expense.
-    """
-    if amount is None or amount == 0:
-        return "0"
-    return "-" if amount < 0 else "+"
 
 
 @dataclass(slots=True)
@@ -226,6 +197,7 @@ class CategorizationService:
         self._audit = audit if audit is not None else AuditService(db)
         self._matcher = CategorizationMatcher(db)
         self._applier = MatchApplier(db, audit=self._audit)
+        self._assist = AssistBridge(db)
 
     # -- Per-transaction category writes (audit emission via applier) --
 
@@ -1187,74 +1159,9 @@ class CategorizationService:
         account_filter: list[str] | None = None,
         date_range: tuple[str, str] | None = None,
     ) -> list[RedactedTransaction]:
-        """Return uncategorized transactions as redacted records for LLM review.
-
-        Sensitivity: medium. Output is sent to the user's LLM via MCP or
-        written to disk via the CLI bridge. The redaction contract is enforced
-        by RedactedTransaction's frozen dataclass shape (v2: description + memo
-        redacted; structural fields exposed unredacted).
-        """
-        import time
-
-        from moneybin.metrics.registry import (
-            CATEGORIZE_ASSIST_DURATION_SECONDS,
-            CATEGORIZE_ASSIST_TXNS_RETURNED_TOTAL,
+        """Return uncategorized transactions as redacted records for LLM review."""
+        return self._assist.categorize_assist(
+            limit=limit,
+            account_filter=account_filter,
+            date_range=date_range,
         )
-
-        settings = get_settings().categorization
-        effective_limit = min(limit, settings.assist_max_batch_size)
-
-        where_clauses = ["tc.transaction_id IS NULL"]
-        params: list[object] = []
-        if account_filter:
-            where_clauses.append(
-                f"t.account_id IN ({','.join('?' * len(account_filter))})"
-            )
-            params.extend(account_filter)
-        if date_range:
-            where_clauses.append("t.transaction_date BETWEEN ? AND ?")
-            params.extend(date_range)
-        where_sql = " AND ".join(where_clauses)
-
-        start = time.monotonic()
-        result: list[RedactedTransaction] = []
-        try:
-            rows = self._db.execute(
-                f"""
-                SELECT t.transaction_id,
-                       t.description,
-                       t.memo,
-                       t.source_type,
-                       t.transaction_type,
-                       t.check_number,
-                       t.is_transfer,
-                       t.transfer_pair_id,
-                       t.payment_channel,
-                       t.amount
-                FROM {FCT_TRANSACTIONS.full_name} t
-                LEFT JOIN {TRANSACTION_CATEGORIES.full_name} tc USING (transaction_id)
-                WHERE {where_sql}
-                LIMIT ?
-                """,  # noqa: S608  # where_sql composed from constants and parameter placeholders
-                params + [effective_limit],
-            ).fetchall()
-
-            result = [
-                RedactedTransaction(
-                    transaction_id=row[0],
-                    description_redacted=redact_for_llm(row[1] or ""),
-                    memo_redacted=redact_for_llm(row[2] or ""),
-                    source_type=row[3] or "",
-                    transaction_type=row[4],
-                    check_number=row[5],
-                    is_transfer=bool(row[6]),
-                    transfer_pair_id=row[7],
-                    payment_channel=row[8],
-                    amount_sign=_amount_sign_label(row[9]),
-                )
-                for row in rows
-            ]
-            return result
-        finally:
-            CATEGORIZE_ASSIST_DURATION_SECONDS.observe(time.monotonic() - start)
-            CATEGORIZE_ASSIST_TXNS_RETURNED_TOTAL.inc(len(result))
